+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- On hover error on certin properties when using linux bindings.
 - Diagnostics when property is missing ';'. Now code does not assume ',' is missing if token is at end of line.
 - Avoid reparsing when onDidChangeContent text content did not change from last state.